+++ libtrig/odo/config.rs
@@ -6,17 +6,19 @@ use core::f64::consts::PI;
 #[derive(Debug, Clone, Copy, PartialEq, Default)]
 pub struct OdometryConfig {
     /// The distance between encoder 1 and encoder 2 in centimeters.
-    pub lateral_wheel_distance: float,
+    pub lateral_wheel_distance: Float64,
     /// The distance between the midpoint of encoder 1 and 2 and encoder 3 in centimeters.
-    pub longitudinal_wheel_distance: float,
+    pub longitudinal_wheel_distance: Float64,
     /// The radius of the wheels in centimeters.
-    pub wheel_radius: float,
+    pub wheel_radius: Float64,
     /// The number of ticks per revolution of the encoders.
-    pub ticks_per_revolution: float,
+    pub ticks_per_revolution: Float64,
 }
 
 impl OdometryConfig {
     /// Creates a new `OdometryConfig` with all fields set to `0.0`.
+    #[inline]
+    #[must_use]
     pub const fn new() -> OdometryConfig {
         OdometryConfig {
             lateral_wheel_distance: 0.0,
@@ -27,25 +29,33 @@ impl OdometryConfig {
     }
 
     /// Modifies `lateral_wheel_distance` and returns `self`.
-    pub const fn set_lateral_wheel_distance(mut self, lateral_wheel_distance: float) -> Self {
+    #[inline]
+    #[must_use]
+    pub const fn set_lateral_wheel_distance(mut self, lateral_wheel_distance: Float64) -> Self {
         self.lateral_wheel_distance = lateral_wheel_distance;
         self
     }
 
     /// Modifies `longitudinal_wheel_distance` and returns `self`.
-    pub const fn set_longitudinal_wheel_distance(mut self, longitudinal_wheel_distance: float) -> Self {
+    #[inline]
+    #[must_use]
+    pub const fn set_longitudinal_wheel_distance(mut self, longitudinal_wheel_distance: Float64) -> Self {
         self.longitudinal_wheel_distance = longitudinal_wheel_distance;
         self
     }
 
     /// Modifies `wheel_radius` and returns `self`.
-    pub const fn set_wheel_radius(mut self, wheel_radius: float) -> Self {
+    #[inline]
+    #[must_use]
+    pub const fn set_wheel_radius(mut self, wheel_radius: Float64) -> Self {
         self.wheel_radius = wheel_radius;
         self
     }
 
     /// Modifies `ticks_per_revolution` and returns `self`.
-    pub const fn set_ticks_per_revolution(mut self, ticks_per_revolution: float) -> Self {
+    #[inline]
+    #[must_use]
+    pub const fn set_ticks_per_revolution(mut self, ticks_per_revolution: Float64) -> Self {
         self.ticks_per_revolution = ticks_per_revolution;
         self
     }
@@ -55,20 +65,25 @@ impl OdometryConfig {
 #[derive(Debug, Clone, Copy, PartialEq, Default)]
 pub(crate) struct OdometryFormulaeConfig {
     /// The distance between encoder 1 and encoder 2 in centimeters.
-    pub lateral_wheel_distance: float,
+    pub lateral_wheel_distance: Float64,
     /// The distance between the midpoint of encoder 1 and 2 and encoder 3 in centimeters.
-    pub longitudinal_wheel_distance: float,
+    pub longitudinal_wheel_distance: Float64,
     /// The number of centimeters per tick.
-    pub cm_per_tick: float,
+    pub cm_per_tick: Float64,
 }
 
 impl OdometryFormulaeConfig {
-    fn cm_per_tick_calc(wheel_radius: float, ticks_per_revolution: float) -> float {
+    /// Calculates the number of centimeters per tick.
+    #[inline]
+    #[must_use]
+    fn cm_per_tick_calc(wheel_radius: Float64, ticks_per_revolution: Float64) -> Float64 {
         wheel_radius * 2.0 * PI / ticks_per_revolution
     }
 }
 
 impl From<OdometryConfig> for OdometryFormulaeConfig {
+    #[inline]
+    #[must_use]
     fn from(config: OdometryConfig) -> Self {
         Self {
             lateral_wheel_distance: config.lateral_wheel_distance,
